refactor: revised calc of most common birth year
The calculation function for the most common birth year had an if clause
to check if there is exactly one most common year of birth or not.
If so, the text message was changed in terms of grammar to fit for
exactly one common year. This is not really necessary as
the else clause can print out one or several most common year of birth.
By eliminating the if clause and just keeping the else clause the
readibility was not really improved but one calculation step was
eliminated to save calculation time.